web-app/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==2.0.1
requests==2.25.1
urllib3==1.26.3
Werkzeug==1.0.1
